Initial version of QueuedWrapperTarget
Only the code is done, nothing else has been added to the NLog project.
Also, I only added it to the .NET 4.0 project.  I'm not familiar with the
innerds of the NLog project yet, so I'm sure there is plenty more to do,
like updating the XML Schema to include this wrapper type in the target
enumeration.